# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "this is the swagger implementation for dialosoft",
        "title": "Dialosoft Swagger API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "@Flussen on github",
            "email": "flussen0@gmail.com"
        },
        "license": {
            "name": "GPL-3.0 License",
            "url": "https://www.gnu.org/licenses/gpl-3.0.html"
        },
        "version": "0.2.0"
    },
    "host": "localhost:8080",
    "basePath": "/dialosoft-api/v1",
    "paths": {
...
        "/management/change-user-role": {
            "post": {
                "security": [
                    {
                        "AccessTokenAuth": []
                    },
                    {
                        "RefreshTokenAuth": []
                    }
                ],
                "description": "Updates the role of a user and invalidates their current refresh token to ensure role changes are applied.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Management"
                ],
                "summary": "Change user role",
                "parameters": [
                    {
                        "description": "Change User Role Data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ChangeUserRole"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "UPDATED - User role updated successfully",
                        "schema": {
                            "$ref": "#/definitions/response.StandardResponse"
                        }
                    },
                    "400": {
                        "description": "BAD REQUEST - Invalid input data or parameters",
                        "schema": {
                            "$ref": "#/definitions/response.StandardError"
                        }
                    },
                    "404": {
                        "description": "NOT FOUND - User or role not found",
                        "schema": {
                            "$ref": "#/definitions/response.StandardError"
                        }
                    },
                    "500": {
                        "description": "INTERNAL SERVER ERROR - Unexpected server error",
                        "schema": {
                            "$ref": "#/definitions/response.StandardError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "request.ChangeUserRole": {
            "type": "object",
            "properties": {
                "roleID": {
                    "type": "string",
                    "example": "8213280e-2000-403a-b375-cdcda6488452"
                },
                "userID": {
                    "type": "string",
                    "example": "8213280e-2000-403a-b375-cdcda6488450"
                }
            }
        },
...
        "response.StandardError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "An error occurred - some context"
                }
            }
        },
        "response.StandardResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "info message"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "AccessTokenAuth": {
            "description": "Bearer token for authentication",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "RefreshTokenAuth": {
            "description": "Refresh token for authentication",
            "type": "apiKey",
            "name": "X-Refresh-Token",
            "in": "header"
        }
    }
}